# public/api/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ESG Hub API",
    "version": "1.0.0",
    "description": "Public API for the Ascent Partners Foundation ESG Knowledge Hub. Access 354 ESG articles and 244 curated external resources covering Environmental, Social, and Governance topics.",
    "contact": {
      "name": "ESG Hub Developers",
      "url": "https://esg-hub.ascent.partners/developers"
    },
    "license": {
      "name": "CC BY-SA 4.0",
      "url": "https://creativecommons.org/licenses/by-sa/4.0/"
    }
  },
  "servers": [
    {
      "url": "https://esg-hub.ascent.partners",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/v1/terms": {
      "get": {
        "summary": "List glossary terms",
        "description": "List terms from the ESG glossary with pagination and optional search.",
        "operationId": "listTerms",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": { "type": "string", "maxLength": 200 },
            "description": "Search within term names"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": { "type": "integer", "default": 0 }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of terms",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": { "type": "array", "items": { "type": "object" } },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer" },
                        "offset": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "has_more": { "type": "boolean" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Propose a new glossary term",
        "description": "Submit a glossary term proposal for human review. Requires Bearer token authorization.",
        "operationId": "proposeTerm",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "definition"],
                "properties": {
                  "name": { "type": "string", "maxLength": 500, "description": "Term name" },
                  "definition": { "type": "string", "maxLength": 10000, "description": "Term definition" },
                  "facets": { "type": "object" },
                  "source_urls": { "type": "array", "items": { "type": "string", "format": "uri" } }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Proposal created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "proposal_id": { "type": "string" },
                    "status": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" },
          "429": { "description": "Rate limited" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "ESG Hub Write Token (required for POST/PATCH endpoints)"
      }
    }
  },
...
}